add `Speaker` object
Like in the prototype, a Speaker instance is the only way to initialise speech.
This implements the `Speaker` object with no async support, everything happens syncronously.
It is done this way because the async api is tricky to implement, so syncronous should suffice for the time being.
For now, the implementation is very basic, containing:
* a new method which initialises a connection with the speech dispatcher daemon, as well as setting the user name and client name to the passed in parameter
* a speak method which should speak the passed in line of text, requires mutable reference to self which could present a problem in the future
* a drop implementation that closes the client explicitly